application/dashboard/requirements.txt
streamlit==1.9.0
pandas
protobuf
mysql-connector-python
matplotlib
numpy